+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE = "scontrini-v17";
+const CACHE = "scontrini-v18";
 const SHELL = ["./", "./index.html", "./manifest.json", "./icon-192.png", "./icon-512.png"];
 
 self.addEventListener("install", (e) => {